+++ src/main/java/com/todoary/ms/src/web/dto/CategoryUpdateRequest.java
@@ -1,6 +1,5 @@
-package com.todoary.ms.src.domain;
+package com.todoary.ms.src.web.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class TodoUpdateCheckRequest {
-    @JsonProperty("isChecked")
-    private boolean isChecked;
+public class CategoryUpdateRequest {
+    private String title;
+    private Integer color;
 }
