+++ src/main/java/com/todoary/ms/src/web/dto/TodoUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.todoary.ms.src.domain;
+package com.todoary.ms.src.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
